+++ src/main/java/com/santalucia/example/api/delegate/DefaultEmployeesApiDelegate.java
@@ -3,11 +3,7 @@ package com.santalucia.example.api.delegate;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.CompletableFuture;
 
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Async;
@@ -17,6 +13,9 @@ import com.santalucia.example.api.model.EmployeeResource;
 import com.santalucia.example.api.server.ListEmployeesApiDelegate;
 import com.santalucia.example.core.mappers.EmployeeDomainMapper;
 import com.santalucia.example.core.service.EmployeeService;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 
 /**
@@ -40,16 +39,16 @@ public class DefaultEmployeesApiDelegate implements ListEmployeesApiDelegate {
    */
   @Async
   @Override
-  public CompletableFuture<ResponseEntity<List<EmployeeResource>>> getEmployeesList(Optional<UUID> xRequestId, Pageable pageable) {
+  public ResponseEntity<List<EmployeeResource>> getEmployeesList(Optional<UUID> xRequestId, Pageable pageable) {
     log.info("Pageable pagenumber: {} ", pageable.getPageNumber());
     log.info("Pageable pageSize: {} ", pageable.getPageSize());
     log.info("Pageable offset: {} ", pageable.getOffset());
 
-    return CompletableFuture.completedFuture(
-        Optional
-          .of(employeeService.getEmployees(pageable))
-          .map(employees -> ResponseEntity.ok()
-          .body(employeeMapper.toResources(employees))) // 200 OK
-          .orElseGet(() -> ResponseEntity.notFound().build())); // 404 Not found
+    return Optional
+            .of(employeeService.getEmployees(pageable))
+            .map(employees -> ResponseEntity.ok()
+            .body(employeeMapper.toResources(employees))) // 200 OK
+            .orElseGet(() -> ResponseEntity.notFound().build()); // 404 Not found
+       
   }
 }