+++ src/main/java/com/santalucia/example/api/delegate/DefaultHelloNameApiDelegate.java
@@ -2,19 +2,21 @@ package com.santalucia.example.api.delegate;
 
 import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.CompletableFuture;
 
-
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import org.springframework.cloud.client.discovery.DiscoveryClient;
+import org.springframework.cloud.client.discovery.composite.CompositeDiscoveryClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
+import org.springframework.web.client.RestClient;
 
 import com.santalucia.example.api.model.IdentidadDigitalConsultaResource;
 import com.santalucia.example.api.server.HelloWorldApiDelegate;
 import com.santalucia.example.core.mappers.IdentidadDigitalDomainMapper;
 import com.santalucia.example.core.service.HelloService;
+
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 
 /**
@@ -28,6 +30,8 @@ public class DefaultHelloNameApiDelegate implements HelloWorldApiDelegate {
 
   private final HelloService helloService;
   private final IdentidadDigitalDomainMapper identidadDigitalDomainMapper;
+  private final RestClient restClient;
+  private final DiscoveryClient discoveryClient;
 
 
   /**
@@ -37,16 +41,27 @@ public class DefaultHelloNameApiDelegate implements HelloWorldApiDelegate {
    */
   @Async
   @Override
-  public CompletableFuture<ResponseEntity<IdentidadDigitalConsultaResource>> getHelloByName(String name, Optional<UUID> xRequestID) {
+  public ResponseEntity<IdentidadDigitalConsultaResource> getHelloByName(String name, Optional<UUID> xRequestID) {
     log.debug("processing getHelloByName");
-
-    return CompletableFuture.completedFuture(
+    
+    CompositeDiscoveryClient compositeDiscoveryClient = (CompositeDiscoveryClient)discoveryClient;
+    
+    log.info("discoveryClient service example-app-11 {}", 
+    		compositeDiscoveryClient.getInstances("example-app-11"));
+   
+    
+    String response = restClient.get().uri("/hello-world/v1/hello/jose").
+    		retrieve().body(String.class);
+    
+    log.info("response {}", response);
+   
+    return 
       Optional
         .ofNullable(helloService.getHelloByName(name))
         .map(idDomain -> ResponseEntity.ok().body(identidadDigitalDomainMapper.toResource(idDomain))) // 200 OK
-        .orElseGet(() -> ResponseEntity.notFound().build()));// 404 Not found
+        .orElseGet(() -> ResponseEntity.notFound().build());// 404 Not found
 
   }
-
-
+  
+  
 }