+++ src/main/java/com/santalucia/example/infrastructure/dto/AccuinsfluDto.java
@@ -1,0 +1,9 @@
+package com.santalucia.example.infrastructure.dto;
+
+public record AccuinsfluDto(
+		String nidinflu,
+		String cflutrab,
+		Integer cactiflu,
+		Double pgradava) {
+
+}